Fix kerf on layout
When kerf was specified, the part was correctly chosen for the board
including a kerf, however, when the part was added to the layout it
didn't include the kerf, thereby reserving less space on the board
than required for the part and allowing other parts to fit in that
space which should not have. This has been fixed for both width and
length.