CutListPlus csv extension to include parts
Per request from John Bonin and with the assistance of
Todd Peterson of CutListPlus, parts are now included in the
csv of the CLP file generated whenever parts option is ticked
on the cutlist options menu.
The format of the csv is per Todd Peterson's suggestions to
John Bonin. John is an apprentice of Joe Zeh of Swamp River Wood Works
in MA.