Changes to csv output for fractions < 1
For both csv files read in by excel or excel like programs
and for CutListPlus, whenever a value is a fraction and less
than 1 (1 unit of whatever), then cutlist displays it in a
format which for excel may be interpreted as a date or for CLP
may be skipped on import.
This change modifies the to_csv utility to remove the ~ (unchanged)
and now also to parse the field being written to the csv to see
if it is in the format ' a b/cUnits'. If it is and a=NULL, b/c
exists and is non-null and Units either exists or else is any string
then a is set to 0.

The format '0 b/cUnits' is satisfatory for both excel and CLP to
recognize this as a fraction rather than text or any other format.