Modify text/background for mac display issues
On some versions of safari, the display shows as black
background with black text, rendering the text fields
invisible. To alleviate this but to ensure compatibility
with all other platforms, the background is changed to
blue and the lettering as white for the 4 text fields
in the gui.
This allows safari users who see only a black background
to at leas see the white lettering. Everyone else will
see a blue background with bold white lettering.
Functionally there is no change with this fix, only a
display change