Modify installation to be compatible with installation menu in SU via .rbz
The installation of the plugin has been made compatible with installation
using the install menu within Sketchup ( Window\Preferences\Extensions )
The plugin is now distributed as a .rbz file.
The .rbz file is simply "ruby zipped" file which is notthing more than a
standard zipped file with a renamed extension to support SU installation
procedure.
The installation can take place either via the Extensions menu or the older
method of unzipping the contents into the Plugins directory.
The advantage of using the new method is that SU takes care of unzipping and
placing the files in the necessary location.