Wrap code in private modules
All code is wrapped in 2 layers of modules to ensure that
all methods are private.
Remove extensions of any Sketchup classes and wrap in
private methods. Modify all references to use new
private methods.
These changes are made to ensure compatibility with other
extensions and to satisfy requirements for inclusion of the
plugin in 'Sketchup Extensions'
The name of the .rbz file is also modified to comply with
Sketchup Extensions standards.
The file is now in the form sr_cutlist_vX.Y.Z.rbz

It is recommended to install the plugin via the Sketchup's extension
installer in the preferences menu, as the unpacking of the
rbz file into the appropriate directories is automatic.