Add sr prefix to cutlist.rb and cutlist directory
The main ruby which adds the CutList plugin is modified
to have prefix 'sr' (my initials) and the same change
is made to the cutlist directory. This is to comply
with changes requested to support placement of the
plugin into Sketchup Extensions. This was at the request
of the code reviewers and gatekeepers at Trimble.
Fussy they are. Though it makes the name unique, I think
it confuses the naming convention because the name is
not obviously the same as the plugin name.
Hopefully this is not important if the plugins are managed
automatically through Sketchup.