Better algo for intl separator characters
A more reliable method of determining the proper decimal character
and csv separators is implemented. Using the locale
determined in Sketchup can default to en-US if the language
of Sketchup doesn't match the language of the OS so this isn't a
reliable method. Instead, a method is implemented which inspects
whether Sketchup uses a decimal or comma for a displayed length
which seems to be a more reliable method since Sketchup must
make a more thorough decision internally. Based on the results
of the length of a test component, the decimal character and
separator is selected.
If the derived locale is English, then "." is useed for decimal
and "," used for csv. Otheriwse "," is used for decimal and ";"
is used as the csv separator.