Hidden dynamic components should be excluded
When a dynamic component is hidden in the selection
then it should not be included. Previously this change
was not required for non dynamic components. Hidden
components do not appear in the selection. This does
not work the same for Dynamic components and an explicit
check for hidden components is added to ensure that it is
not included in the materials list.