Correct decimal notation derivation with groups
When the model contains primarily groups or the
in particular when the first part is made from
a group definition, then the decimal character
calculation can fail and CutList will produce
no output. This has been fixed. This problem
only exists with v4.1.9 so a workaround is
use v4.1.8 or to convert your groups to
components or (now) to download v4.1.10.