Correct display digits rounding for some fields
The total length field and total board feet for
compact list summaries sometimes displays a
long float value with no rounding. This was showing up only in SKU2014
Beta and not previus versions of SKU. This is being corrected for
release of SKU2014.

Upgrade method of upgrading version so only a single location needs to
be changed each time.

Upgrade version to v4.1.11