Part size comparison for layout sometimes fails
Internally a part size may become a float value
because of some scaling and calculations done.
This may result in incorrect comparisons when
trying to place the part on a board. The board
might fail to place because of minor discrepancies
between actual size and computed values even
though the part fits within the accuracy of the
model unit scale selected. The comparisons are
now done to the accuracy of the model.