Replace .typename with more efficient .is_a?
Entity types in Sketchup can be face, group, componentinstance
etc. The API publishes the recommended way of distinguishing
them by method .typename. However typename is a string and a
string comparison is relatively slow so a faster way is to
to test it with the .is_a? against Sketchup::Face,
Sketchup::Group, Sketchup::ComponentInstance for example.
This method has been show to be about 80% more efficient
and is the recommended method (though though not readily
appearing in the documentation).